+++ examples/resources/sdwan_system_mrf_feature/import.sh
@@ -1,3 +1,2 @@
-
 # Expected import identifier with the format: "system_mrf_feature_id,feature_profile_id"
 terraform import sdwan_system_mrf_feature.example "f6b2c44c-693c-4763-b010-895aa3d236bd,f6dd22c8-0b4f-496c-9a0b-6813d1f8b8ac"