+++ examples/resources/sdwan_transport_wan_vpn_interface_gre_feature_associate_tracker_feature/import.sh
@@ -1,1 +1,2 @@
-terraform import sdwan_transport_wan_vpn_interface_gre_feature_associate_tracker_feature.example "f6b2c44c-693c-4763-b010-895aa3d236bd"
+# Expected import identifier with the format: "transport_wan_vpn_interface_gre_feature_associate_tracker_feature_id,feature_profile_id,transport_wan_vpn_feature_id,transport_wan_vpn_interface_gre_feature_id"
+terraform import sdwan_transport_wan_vpn_interface_gre_feature_associate_tracker_feature.example "f6b2c44c-693c-4763-b010-895aa3d236bd,f6dd22c8-0b4f-496c-9a0b-6813d1f8b8ac,140331f6-5418-4755-a059-13c77eb96037,140331f6-5418-4755-a059-13c77eb96037"