+++ internal/provider/model_sdwan_transport_wan_vpn_feature_associate_routing_bgp_feature.go
@@ -33,6 +33,7 @@ import (
 // Section below is generated&owned by "gen/generator.go". //template:begin types
 type TransportWANVPNFeatureAssociateRoutingBGPFeature struct {
 	Id                           types.String `tfsdk:"id"`
+	Version                      types.Int64  `tfsdk:"version"`
 	FeatureProfileId             types.String `tfsdk:"feature_profile_id"`
 	TransportWanVpnFeatureId     types.String `tfsdk:"transport_wan_vpn_feature_id"`
 	TransportRoutingBgpFeatureId types.String `tfsdk:"transport_routing_bgp_feature_id"`